[MODFIY]: AddPage0 ViewModel 분리
