+++ lib/pages/home_page/home_page_view_model/home_page_view_model.dart
@@ -39,7 +39,6 @@ class HomePageViewModel extends ChangeNotifier {
     Navigator.of(context).push(
       CupertinoPageRoute(
         builder: (_) => AddPage_0(
-          accessToken: accessToken,
           sigun: sigunguText(context),
         ),
       ),
