+++ lib/pages/home_page/widgets/upcoming_plan_section.dart
@@ -72,7 +72,7 @@ class UpcomingPlanSection extends StatelessWidget {
         Navigator.push(
           context,
           CupertinoPageRoute(
-            builder: (_) => AddPage_0(accessToken: accessToken),
+            builder: (_) => AddPage_0(),
           ),
         );
       },
