+++ lib/provider.dart
@@ -11,11 +11,13 @@ import 'package:alpha_fe/services/websocket/show_tour_course/show_tour_course_we
 import 'package:provider/provider.dart';
 import 'package:provider/single_child_widget.dart';
 import 'package:alpha_fe/pages/home_page/home_page_view_model/home_page_view_model.dart';
+import 'package:alpha_fe/pages/add_page/add_page_0/view_model/add_page_0_view_model.dart';
 
 
 
 List<SingleChildWidget> appProviders = [
   ChangeNotifierProvider(create: (_) => HomePageViewModel()),
+  ChangeNotifierProvider(create: (_) => AddPage0ViewModel()),
   ChangeNotifierProvider(create: (_) => AuthProvider()),
   ChangeNotifierProvider(create: (_) => PlanViewModel()),
   ChangeNotifierProvider(create: (_) => LoadingViewModel()),
